feat(config): make additional test dependencies configurable
The shared tasks come with a default set of defined glob patterns to
specify files for the karma test runner to be loaded. However, sometimes
a sofa component has more dependencies than the onces that are pre-defined
so we need a hook to let components configure, which additional dependencies
are needed for tests.

It's now possible to define an array of file paths that point to additional
dependencies by providing a `testDependencyFiles` member on the components
configuration object.

Example:

```js
require('shared-sofa-component-tasks')(require('gulp'), {
  pkg: require('./package.json'),
  baseDir: __dirname,
  testDependencyFiles: [
    'node_modules/sofa-tracking/dist/sofa.tracking.js'
  ]
});
```