feat(deployTask): adds deploy task for easy releasing
In your component you can now run `gulp deploy` to release a new
version of a component. The task does the following things:

- Bumps `package.json` and `bower.json` version
- Creates/Updates `CHANGELOG.md`
- Generates a distribution
- Creates a release commit
- Pushes release incl. generated Tag to GitHub
- Publishes a new version of the corresponding npm module

In order to specify the version release type, you can add one of the
following types as option in the command line

- prerelease
- patch
- minor
- major

So e.g. to make a major release run `gulp deploy --major`